segundo taller/grpc-prueba-master/pyproject.toml
[tool.poetry]
name = "grpc"
version = "0.1.0"
description = ""
authors = ["Santicm23 <Santicastrom@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.62.0"
grpcio-tools = "^1.62.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
